Move web view test options setup to teset app
