+++ Adjust/sdk-plugin-webbridge/src/main/java/com/adjust/sdk/webbridge/AdjustBridgeInstance.java
@@ -672,116 +672,6 @@ public class AdjustBridgeInstance {
     }
 
     @JavascriptInterface
-    public void setTestOptions(final String testOptionsString) {
-        AdjustFactory.getLogger().verbose("AdjustBridgeInstance setTestOptions: %s", testOptionsString);
-
-        if (!isInitialized()) {
-            return;
-        }
-
-        try {
-            AdjustTestOptions adjustTestOptions = new AdjustTestOptions();
-            JSONObject jsonAdjustTestOptions = new JSONObject(testOptionsString);
-
-            Object baseUrlField = jsonAdjustTestOptions.get("baseUrl");
-            Object gdprUrlField = jsonAdjustTestOptions.get("gdprUrl");
-            Object subscriptionUrlField = jsonAdjustTestOptions.get("subscriptionUrl");
-            Object basePathField = jsonAdjustTestOptions.get("basePath");
-            Object gdprPathField = jsonAdjustTestOptions.get("gdprPath");
-            Object subscriptionPathField = jsonAdjustTestOptions.get("subscriptionPath");
-            Object useTestConnectionOptionsField = jsonAdjustTestOptions.get("useTestConnectionOptions");
-            Object connectionOptionsField = jsonAdjustTestOptions.get("useTestConnectionOptions");
-            Object timerIntervalInMillisecondsField = jsonAdjustTestOptions.get("timerIntervalInMilliseconds");
-            Object timerStartInMillisecondsField = jsonAdjustTestOptions.get("timerStartInMilliseconds");
-            Object sessionIntervalInMillisecondsField = jsonAdjustTestOptions.get("sessionIntervalInMilliseconds");
-            Object subsessionIntervalInMillisecondsField = jsonAdjustTestOptions.get("subsessionIntervalInMilliseconds");
-            Object teardownField = jsonAdjustTestOptions.get("teardown");
-            Object tryInstallReferrerField = jsonAdjustTestOptions.get("tryInstallReferrer");
-            Object noBackoffWaitField = jsonAdjustTestOptions.get("noBackoffWait");
-            Object hasContextField = jsonAdjustTestOptions.get("hasContext");
-
-            String gdprUrl = AdjustBridgeUtil.fieldToString(gdprUrlField);
-            if (gdprUrl != null) {
-                adjustTestOptions.gdprUrl = gdprUrl;
-            }
-
-            String baseUrl = AdjustBridgeUtil.fieldToString(baseUrlField);
-            if (baseUrl != null) {
-                adjustTestOptions.baseUrl = baseUrl;
-            }
-
-            String subscriptionUrl = AdjustBridgeUtil.fieldToString(subscriptionUrlField);
-            if (subscriptionUrl != null) {
-                adjustTestOptions.subscriptionUrl = subscriptionUrl;
-            }
-
-            String basePath = AdjustBridgeUtil.fieldToString(basePathField);
-            if (basePath != null) {
-                adjustTestOptions.basePath = basePath;
-            }
-
-            String gdprPath = AdjustBridgeUtil.fieldToString(gdprPathField);
-            if (gdprPath != null) {
-                adjustTestOptions.gdprPath = gdprPath;
-            }
-
-            String subscriptionPath = AdjustBridgeUtil.fieldToString(subscriptionPathField);
-            if (subscriptionPath != null) {
-                adjustTestOptions.subscriptionPath = subscriptionPath;
-            }
-
-            Boolean useTestConnectionOptions = AdjustBridgeUtil.fieldToBoolean(useTestConnectionOptionsField);
-            if (useTestConnectionOptions != null) {
-                adjustTestOptions.connectionOptions = AdjustBridgeUtil.testConnectionOptions();
-            }
-
-            Long timerIntervalInMilliseconds = AdjustBridgeUtil.fieldToLong(timerIntervalInMillisecondsField);
-            if (timerIntervalInMilliseconds != null) {
-                adjustTestOptions.timerIntervalInMilliseconds = timerIntervalInMilliseconds;
-            }
-
-            Long timerStartInMilliseconds = AdjustBridgeUtil.fieldToLong(timerStartInMillisecondsField);
-            if (timerStartInMilliseconds != null) {
-                adjustTestOptions.timerStartInMilliseconds = timerStartInMilliseconds;
-            }
-
-            Long sessionIntervalInMilliseconds = AdjustBridgeUtil.fieldToLong(sessionIntervalInMillisecondsField);
-            if (sessionIntervalInMilliseconds != null) {
-                adjustTestOptions.sessionIntervalInMilliseconds = sessionIntervalInMilliseconds;
-            }
-
-            Long subsessionIntervalInMilliseconds = AdjustBridgeUtil.fieldToLong(subsessionIntervalInMillisecondsField);
-            if (subsessionIntervalInMilliseconds != null) {
-                adjustTestOptions.subsessionIntervalInMilliseconds = subsessionIntervalInMilliseconds;
-            }
-
-            Boolean teardown = AdjustBridgeUtil.fieldToBoolean(teardownField);
-            if (teardown != null) {
-                adjustTestOptions.teardown = teardown;
-            }
-
-            Boolean tryInstallReferrer = AdjustBridgeUtil.fieldToBoolean(tryInstallReferrerField);
-            if (tryInstallReferrer != null) {
-                adjustTestOptions.tryInstallReferrer = tryInstallReferrer;
-            }
-
-            Boolean noBackoffWait = AdjustBridgeUtil.fieldToBoolean(noBackoffWaitField);
-            if (noBackoffWait != null) {
-                adjustTestOptions.noBackoffWait = noBackoffWait;
-            }
-
-            Boolean hasContext = AdjustBridgeUtil.fieldToBoolean(hasContextField);
-            if (hasContext != null && hasContext.booleanValue()) {
-                adjustTestOptions.context = application.getApplicationContext();
-            }
-
-            Adjust.setTestOptions(adjustTestOptions);
-        } catch (Exception e) {
-            AdjustFactory.getLogger().error("AdjustBridgeInstance setTestOptions: %s", e.getMessage());
-        }
-    }
-
-    @JavascriptInterface
     public void fbPixelEvent(String pixelId, String event_name, String jsonString) {
         this.facebookSDKJSInterface.sendEvent(pixelId, event_name, jsonString);
     }
